add edit query widget
shared between create/edit query dialogs
