+++ src/admc/CMakeLists.txt
@@ -51,6 +51,7 @@ set(ADMC_SOURCES
     policy_results_widget.cpp
     gplink.cpp
     filter_classes_widget.cpp
+    edit_query_widget.cpp
     
     console_types/console_object.cpp
     console_types/console_policy.cpp
